requirements.txt
torch~=2.0.1
tqdm~=4.65.0
transformers~=4.28.0.dev0
numpy~=1.24.3
Pillow~=9.5.0
decord~=0.6.0
gradio~=3.23.0
markdown2~=2.4.8
einops~=0.6.1
requests~=2.30.0